requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
audio-recorder-streamlit==0.0.10
av==12.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
dnspython==2.6.1
entrypoints==0.4
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
ifaddr==0.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.6.2
numpy==2.1.1
openai==1.44.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
SpeechRecognition==3.10.4
streamlit==1.38.0
streamlit-audiorec==0.1.3
streamlit-audiorecorder==0.0.5
streamlit-float==0.3.5
streamlit-webrtc==0.47.7
streamlit_mic_recorder==0.0.8
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
